backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.45.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "deptry>=0.24.0",
    "httpx>=0.28.1",
    "pre-commit>=4.5.0",
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.9",
]

[tool.deptry.per_rule_ignores]
DEP002 = ["uvicorn"]
